T#799 P3-D: Extract remote module from server.ts
Sub-task D of T#786 Phase 3. Extract 3 remote tmux-window-attach handlers
(~90 LoC) + REMOTE_SESSION constant + attachedBeastName state from server.ts
into src/remote/routes.ts.

## Changes

**New: src/remote/routes.ts**
- `registerRemoteRoutes(app, helpers)` registrar
- RemoteHelpers DI: `isLocalNetwork`, `hasSessionAuth`
- Module-local `REMOTE_SESSION` constant + `attachedBeastName` state (closure-scoped, resets on server restart — tmux session is source of truth)
- 3 handlers extracted verbatim:
  - GET /api/remote/status — current attached beast (with tmux liveness check)
  - POST /api/remote/attach — link a beast's claude tmux window to Mindlink session (local-only)
  - POST /api/remote/detach — unlink current beast

**Modified: src/server.ts**
- Add `import { registerRemoteRoutes } from './remote/routes.ts'`
- Remove `REMOTE_SESSION` constant + `attachedBeastName` mutable state
- Replace 3 inline handler blocks (~92 lines) with single `registerRemoteRoutes(app, { isLocalNetwork, hasSessionAuth })`
- server.ts line count: 4320 → 4228 (-92 lines net)

`execSync` import in server.ts is preserved (still used by beast-terminal endpoints at lines 2217/2225/2240, separate concern).

## Test plan

- [x] `bun build src/server.ts --target=bun --outdir=/tmp/t799-build` clean — all imports resolve at bundle-time (per the Norm #84 refinement pattern Zaghnal acked at T#798 #1135).
- [x] DI captures all closure deps explicitly (`isLocalNetwork`, `hasSessionAuth`).
- [x] REMOTE_SESSION moved into the module that uses it (no orphaned constant in server.ts).
- [ ] Post-deploy smoke at deploy gate:
  - `GET /api/remote/status` → 200 with `{session_exists, attached_beast}` shape (no-auth read OK; module-local state)
  - `POST /api/remote/attach` with `{beast: "invalid"}` → 400 "Invalid beast name"
  - `POST /api/remote/attach` from non-local without session → 403 "forbidden"

Reviewer: @zaghnal per task assignment.

## References

- T#799: sub-task of T#786 P3 (sibling under T#764 per Spec #56 v1)
- T#798: established the bundle-build + post-deploy smoke pattern Zaghnal acked
- T#783: module-registration-must-precede-SPA-catch-all rule respected (registration at line 2340 < SPA at 4119)
