+++ src/server.ts
@@ -93,6 +93,7 @@ import type { Role } from './server/rbac.ts';
 import { registerGovernanceRoutes } from './governance/routes.ts';
 import { registerProwlRoutes } from './prowl/routes.ts';
 import { registerSettingsRoutes } from './settings/routes.ts';
+import { registerRemoteRoutes } from './remote/routes.ts';
 import { registerLibraryRoutes } from './library/routes.ts';
 import { registerRiskRoutes } from './risk/routes.ts';
 import { registerTelegramRoutes } from './telegram/routes.ts';
@@ -2336,101 +2337,8 @@ function getTmuxStatus(): { tmuxStatus: Map<string, 'processing' | 'idle' | 'wai
 // Remote Control — tmux Beast switcher
 // ============================================================================
 
-const REMOTE_SESSION = 'Mindlink';
-let attachedBeastName: string | null = null;
-
-// GET /api/remote/status — which beast is currently attached
-app.get('/api/remote/status', (c) => {
-  // Verify the Remote session still exists and has a linked window
-  if (attachedBeastName) {
-    try {
-      execSync(`tmux has-session -t ${JSON.stringify(REMOTE_SESSION)}`, { timeout: 2000 });
-      // Check if window 1 still exists (beast is still linked)
-      const windows = execSync(
-        `tmux list-windows -t ${JSON.stringify(REMOTE_SESSION)} -F "#{window_index}"`,
-        { timeout: 2000 }
-      ).toString().trim().split('\n');
-      if (!windows.includes('1')) {
-        attachedBeastName = null; // Window was unlinked externally
-      }
-    } catch {
-      attachedBeastName = null; // Session gone
-    }
-  }
-
-  return c.json({ session_exists: !!attachedBeastName, attached_beast: attachedBeastName });
-});
-
-// POST /api/remote/attach — attach a beast's claude window (local only — requires tmux)
-app.post('/api/remote/attach', async (c) => {
-  // Remote attach requires local tmux access — reject non-local requests cleanly
-  if (!isLocalNetwork(c) && !hasSessionAuth(c)) {
-    return c.json({ error: 'forbidden' }, 403);
-  }
-  try {
-    const data = await c.req.json();
-    const beastName = data.beast?.toLowerCase();
-    if (!beastName) return c.json({ error: 'beast name required' }, 400);
-
-    // Sanitize: only allow alphanumeric beast names
-    if (!/^[a-z]+$/.test(beastName)) return c.json({ error: 'Invalid beast name' }, 400);
-
-    const sessionName = beastName.charAt(0).toUpperCase() + beastName.slice(1);
-
-    // Verify beast session exists
-    try {
-      execSync(`tmux has-session -t ${JSON.stringify(sessionName)}`, { timeout: 2000 });
-    } catch {
-      return c.json({ error: `No tmux session for ${beastName}` }, 404);
-    }
-
-    // Find the claude window index in the beast's session
-    let claudeWindow = '1';
-    try {
-      const windows = execSync(
-        `tmux list-windows -t ${JSON.stringify(sessionName)} -F "#{window_index}:#{pane_current_command}"`,
-        { timeout: 2000 }
-      ).toString().trim().split('\n');
-      const claudeWin = windows.find(w => w.includes(':claude'));
-      if (claudeWin) claudeWindow = claudeWin.split(':')[0];
-    } catch { /* default to 1 */ }
-
-    // Ensure Remote session exists
-    try {
-      execSync(`tmux has-session -t ${JSON.stringify(REMOTE_SESSION)}`, { timeout: 2000 });
-    } catch {
-      execSync(`tmux new-session -d -s ${JSON.stringify(REMOTE_SESSION)}`, { timeout: 2000 });
-    }
-
-    // Unlink any existing beast window (window index 1)
-    try {
-      execSync(`tmux unlink-window -k -t ${JSON.stringify(REMOTE_SESSION)}:1`, { timeout: 2000 });
-    } catch { /* no window to unlink */ }
-
-    // Link the beast's claude window
-    execSync(
-      `tmux link-window -s ${JSON.stringify(sessionName)}:${claudeWindow} -t ${JSON.stringify(REMOTE_SESSION)}:1`,
-      { timeout: 2000 }
-    );
-
-    // Switch to the linked window
-    execSync(`tmux select-window -t ${JSON.stringify(REMOTE_SESSION)}:1`, { timeout: 2000 });
-
-    attachedBeastName = beastName;
-    return c.json({ attached: beastName, session: REMOTE_SESSION });
-  } catch (e) {
-    return c.json({ error: e instanceof Error ? e.message : 'Attach failed' }, 500);
-  }
-});
-
-// POST /api/remote/detach — detach current beast (local only — requires tmux)
-app.post('/api/remote/detach', (_c) => {
-  try {
-    execSync(`tmux unlink-window -k -t ${JSON.stringify(REMOTE_SESSION)}:1`, { timeout: 2000 });
-  } catch { /* already detached */ }
-  attachedBeastName = null;
-  return _c.json({ detached: true });
-});
+// Remote routes extracted to src/remote/routes.ts (T#799 P3-D)
+registerRemoteRoutes(app, { isLocalNetwork, hasSessionAuth });
 
 // ============================================================================
 // Beast Profile Routes
